assets/requirements.txt
# DjAdmin v1.0.20240424 
# https://github.com/kanmain/djadmin

# Django
django
djangorestframework

# Environment
python-decouple

# DATABASES
# Mysql in ubuntu: sudo apt-get install python3-dev default-libmysqlclient-dev build-essential pkg-config
mysqlclient
# psycopg2 for production
psycopg2-binary 
